Move fields code into `proof::field.rs`
