+++ ledger/src/proofs/mod.rs
@@ -4,6 +4,7 @@ pub mod accumulator_check;
 pub mod block;
 pub mod caching;
 pub mod constants;
+pub mod field;
 pub mod group_map;
 pub mod merge;
 pub mod numbers;
